Change scheduleRepeatingTask() to scheduleTask()
rejoinTimerTask runs rejoinUntilConnected() which checks for the current channel\socket state and either rejoin channel or schedule the next check. So the task must be set to start only once after each check.
Current behaviour leads to the memory allocation error and application crash in case of the network problems.